chore: split service folder
