+++ wallets-testing/pages/widget.page.ts
