+++ wallets-testing/test/widgets/ethereum.spec.ts
@@ -10,11 +10,7 @@ import {
   METAMASK_COMMON_CONFIG,
 } from '@lidofinance/wallets-testing-wallets';
 import { test } from '@playwright/test';
-import {
-  connectWallet,
-  initBrowserWithExtension,
-  stake,
-} from '../../services/service.test';
+import { connectWallet, initBrowserWithExtension, stake } from '../../services';
 import { BrowserService } from '@lidofinance/browser-service';
 
 test.describe('Ethereum', () => {
